+++ extensions/procfs/processors/ProcFsMonitor.h
@@ -23,7 +23,7 @@
 #include <vector>
 #include <optional>
 
-#include "ProcFs.h"
+#include "../ProcFs.h"
 #include "core/Processor.h"
 #include "utils/Enum.h"
